migrate Emoji to folder
