+++ cogs/emoji/__init__.py
@@ -1,0 +1,5 @@
+from .cog import Emoji
+
+
+def setup(bot):
+    bot.add_cog(Emoji(bot))
