+++ cogs/emoji/messages_cz.py
@@ -1,0 +1,7 @@
+from config.messages import Messages as GlobalMessages
+
+
+class MessagesCZ(GlobalMessages):
+    emoji_get_emojis_brief = "Pošle zip se všemi emoji a stickery ze serveru"
+    emoji_get_emoji_brief = "Zobrazí emoji v plné velikosti"
+    emoji_not_emoji = "Neplatný formát emoji"
